+++ extension.js
@@ -29,11 +29,12 @@ class MatlabFormatter {
 
     format(document, range) {
         return new Promise((resolve, reject) => {
-            let formatter = vscode.workspace.getConfiguration('matlab-formatter')['path'];
+            let formatter = this.py + vscode.workspace.getConfiguration('matlab-formatter')['path'];
+            let indentwidth = " " + vscode.workspace.getConfiguration('matlab-formatter')['indentwidth'];
             let filename = ' "' + document.fileName + '"';
             let start = " " + (range.start.line + 1);
             let end = " " + (range.end.line + 1);
-            cp.exec(this.py + formatter + filename + start + end, (err, stdout, stderr) => {
+            cp.exec(formatter + filename + indentwidth + start + end, (err, stdout, stderr) => {
                 if (stdout != '') {
                     let toreplace = document.validateRange(new vscode.Range(range.start.line, 0, range.end.line+1, 0));
                     var edit = [vscode.TextEdit.replace(toreplace, stdout)];