+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to the "matlab-formatter" extension will be documented in th
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ### 1.1.0
-Add feature: format selection
+New Features:
+* format selection
+* variable indent width → setting: `matlab-formatter.indentwidth`
+
 
 ### 1.0.1